+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Change Log
 ==========
+### Stage
+#### Features
+* adiciona deleção das checklists (PL-3217)
 
 <a name="v1.3.4"></a>
 ## Version [v1.3.4](https://github.com/luizfp/PrologWebService/compare/v1.3.3...v1.3.4) (release-date) [unreleased]